test: Split out Handoff Orchestration tests
